% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWeatherData.R
\name{GetCurrentWeather}
\alias{GetCurrentWeather}
\title{Get Current Weather Data}
\usage{
GetCurrentWeather(location, days = NULL, ...)
}
\arguments{
\item{location}{A character string representing the location for which you want
to retrieve the weather data. This can be a city name, coordinates (latitude and longitude),
or a postal code.}

\item{days}{Optional. Number of Days to supply forecast. If supplied returned
forecasted weather for those days ahead (an positive integer up to 3)}
}
\value{
A list containing the current weather data for the specified location.
If the request fails, it returns an error message.
}
\description{
This function retrieves the current weather data for a specified
location using the WeatherAPI.
}
\examples{
\dontrun{
location <- "London"
current_weather <- GetCurrentWeather(location)
print(current_weather)
}
}
\author{
Gareth Burns
}
