% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsItShortsWeather.R
\name{IsItShortsWeather}
\alias{IsItShortsWeather}
\title{Should I Wear Shorts Today?}
\usage{
IsItShortsWeather(location)
}
\arguments{
\item{location}{A character string representing the location for which you want
to retrieve the weather data. This can be a city name, coordinates (latitude and longitude),
or a postal code.}
}
\value{
A answer to the question 'Is it shorts weather?" for the supplied location.
}
\description{
Tells me if I should wear short in a specific location today.
}
\examples{
\dontrun{
IsItShortsWeather(location = "Belfast")
}
}
\author{
Gareth Burns
}
