% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShouldIDryMyWashing.R
\name{ShouldIDryMyWashing}
\alias{ShouldIDryMyWashing}
\title{Should I Dry My Washing?}
\usage{
ShouldIDryMyWashing(location)
}
\arguments{
\item{location}{A character string representing the location for which you want
to retrieve the weather data. This can be a city name, coordinates (latitude and longitude),
or a postal code.}
}
\value{
A answer to the question 'Should I dry my washing?" for the supplied location.
}
\description{
Tells me should I dry my washing in a location today.
}
\examples{
\dontrun{
ShouldIDryMyWashing(location = "Belfast")
}
}
\author{
Gareth Burns
}
